% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentSet-class.R,
%   R/EnrichmentSet-methods.R
\name{EnrichmentSet}
\alias{EnrichmentSet}
\alias{show,EnrichmentSet-method}
\alias{summary,EnrichmentSet-method}
\title{Constructor for EnrichmentSet}
\usage{
EnrichmentSet(data, metadata, sep = ";")

\S4method{show}{EnrichmentSet}(object)

\S4method{summary}{EnrichmentSet}(object)
}
\arguments{
\item{data}{A data.frame containing at least columns set_id, set_name, feature_ids.}

\item{metadata}{A named list with mapping metadata.}

\item{sep}{Character separator used in feature_ids (default = ";").}

\item{object}{An \code{EnrichmentSet} object.}
}
\value{
An object of class EnrichmentSet.
}
\description{
Constructor for EnrichmentSet
}
