% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentSet-methods.R
\name{EnrichmentSet_as_list}
\alias{EnrichmentSet_as_list}
\title{Coerce EnrichmentSet to list}
\description{
Converts an EnrichmentSet object to a named list of feature vectors,
suitable for enrichment analysis.
}
