% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sets_by_features.R
\name{filter_sets_by_features}
\alias{filter_sets_by_features}
\title{Filter an EnrichmentSet by feature overlap}
\usage{
filter_sets_by_features(eset, features, min_overlap = 1)
}
\arguments{
\item{eset}{An \code{EnrichmentSet} object.}

\item{features}{Character vector of feature IDs to check.}

\item{min_overlap}{Minimum number of overlapping features (default = 1).}
}
\value{
A reduced \code{EnrichmentSet}.
}
\description{
Keep only sets with at least \code{min_overlap} features present in \code{features}.
}
