% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentSet-utils.R
\name{filterEnrichmentSet}
\alias{filterEnrichmentSet}
\title{Filter EnrichmentSet by feature IDs}
\usage{
filterEnrichmentSet(Eset, ids)
}
\arguments{
\item{Eset}{An object of class \code{EnrichmentSet}}

\item{ids}{Character vector of feature IDs to retain}
}
\value{
A filtered \code{EnrichmentSet} object
}
\description{
Returns a new EnrichmentSet containing only sets that include
at least one of the provided feature IDs.
}
