% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot enrichment results}
\usage{
plot_enrichment(
  results,
  type = c("bar", "dot"),
  top = 20,
  x_measure = "-log10(p_adj)",
  color_by = "p_adj",
  title = NULL
)
}
\arguments{
\item{results}{A tibble produced by \code{ora_test()} or \code{qea_test()}.}

\item{type}{Type of plot: \code{"bar"} (default) or \code{"dot"}.}

\item{top}{Integer: number of top sets to display (default = 20).}

\item{x_measure}{Variable for x-axis when type = "bar":
one of \code{"-log10(p_adj)"} (default), \code{"-log10(p_value)"}, or
\code{"fold_enrichment"}.}

\item{color_by}{Variable for coloring points when \code{type = "dot"}:
one of \code{"p_adj"} (default), \code{"p_value"}, or \code{"fold_enrichment"}.}

\item{title}{Optional plot title.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize results from enrichment analyses such as ORA or QEA.
}
