% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentSet-class.R
\name{validateEnrichmentSet}
\alias{validateEnrichmentSet}
\title{Validate EnrichmentSet structure}
\usage{
validateEnrichmentSet(Eset)
}
\arguments{
\item{Eset}{An object of class \code{EnrichmentSet}.}
}
\value{
Invisibly returns TRUE if valid, otherwise throws an error.
}
\description{
Checks that an object is a valid \code{EnrichmentSet} and contains the required columns.
}
