% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentSet-class.R
\name{EnrichmentSet}
\alias{EnrichmentSet}
\alias{show,EnrichmentSet-method}
\alias{summary,EnrichmentSet-method}
\alias{EnrichmentSet-coercion}
\title{Constructor for EnrichmentSet}
\usage{
EnrichmentSet(data, metadata, sep = ";")

\S4method{show}{EnrichmentSet}(object)

\S4method{summary}{EnrichmentSet}(object)
}
\arguments{
\item{data}{A data.frame containing at least columns set_id, set_name, feature_ids.}

\item{metadata}{A named list with mapping metadata.}

\item{sep}{Character separator used in feature_ids (default = ";").}
}
\value{
An object of class EnrichmentSet.
}
\description{
Converts an EnrichmentSet object to a named list of feature vectors,
suitable for enrichment analysis.
}
