% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentSet-methods.R
\name{EnrichmentSet_as_MetaboliteSetDataFrame}
\alias{EnrichmentSet_as_MetaboliteSetDataFrame}
\alias{as.MetaboliteSetDataFrame}
\title{Coerce EnrichmentSet to a condensed metabolite-set table}
\usage{
as.MetaboliteSetDataFrame(x, id_sep = ",")
}
\arguments{
\item{x}{An \code{EnrichmentSet} object.}

\item{id_sep}{Separator used to concatenate feature IDs (default = ",").}
}
\value{
A data.frame with columns \code{set_name} and \code{Metabolites}.
}
\description{
Converts an EnrichmentSet object to a two-column data.frame
with one row per set and a comma-separated string of metabolite IDs.
}
